2019-10-22 M:N data

+++ 03_django/03_django_form/articles/views.py
@@ -1,3 +1,4 @@
+import hashlib
 from IPython import embed
 from django.contrib.auth.decorators import login_required
 from django.shortcuts import render, redirect, get_object_or_404
@@ -12,6 +13,11 @@ from .forms import ArticleForm, CommentForm
 # Create your views here.
 
 def index(request):
+    # if request.user.is_authenticated:
+    #     gravatar_url = hashlib.md5(request.user.email.encode('utf-8').lower().strip()).hexdigest()
+    # else:
+    #     gravatar_url = None
+    # 필터를 만듦으로써 필터의 중복을 처리했다.
     # session에 visits_num 키로 접근해 값을 가져온다.
     # 기본적으로 존재하지 않는 키이기 때문에 키가 없다면(방문한적이 없다면) 0 값을 가져오도록 한다.
     visits_num = request.session.get('visits_num', 0)
@@ -32,7 +38,9 @@ def create(request):
         # form 이 유효한지 체크한다.
         if form.is_valid():
             #form.cleaned_data 로 정제된 데이터를 받는다.
-            article = form.save()
+            article = form.save(commit=False)
+            article.user = request.user
+            article.save()
             return redirect(article)
     else:
         form = ArticleForm()
@@ -54,23 +62,29 @@ def detail(request, article_pk):
 def delete(request, article_pk):
     if request.user.is_authenticated:
         article = get_object_or_404(Article, pk=article_pk)
-        article.delete()
+        if request.user == article.user:
+            article.delete()
+        else:
+            return redirect(article)
     return redirect('articles:index')
 
 @login_required
 def update(request, article_pk):
     article = get_object_or_404(Article, pk=article_pk)
-    if request.method == 'POST':
-        form = ArticleForm(request.POST, instance=article)
-        if form.is_valid():            
-            article = form.save()
-            return redirect(article)
+    if request.user == article.user:
+        if request.method == 'POST':
+            form = ArticleForm(request.POST, instance=article)
+            if form.is_valid():            
+                article = form.save()
+                return redirect(article)
+        else:
+            # ArticleForm 을 초기화(이전에 DB에 저장된 데이터를 넣어준 상태)
+            # form = ArticleForm(initial={'title':article.title, 'content':article.content})
+            # __dict__ : article 객체 데이터를 딕셔너리 자료형으로 변환
+            # form = ArticleForm(initial=article.__dict__)
+            form = ArticleForm(instance=article)
     else:
-        # ArticleForm 을 초기화(이전에 DB에 저장된 데이터를 넣어준 상태)
-        # form = ArticleForm(initial={'title':article.title, 'content':article.content})
-        # __dict__ : article 객체 데이터를 딕셔너리 자료형으로 변환
-        # form = ArticleForm(initial=article.__dict__)
-        form = ArticleForm(instance=article)
+        return redirect('articles:index')
     # 1. POST: 검증에 실패한 form(오류 메세지도 포험된 상태)
     # 2. GET : 초기화된 forms
     context = {'form':form, 'article':article,}
@@ -84,6 +98,7 @@ def create_comment(request, article_pk):
             #객체를 Create 하지만, db에 레코드는 작성하지 않는다.
             comment = comment_form.save(commit=False)
             comment.article_id = article_pk
+            comment.user = request.user
             comment.save()
     return redirect('articles:detail', article_pk)
 
@@ -91,6 +106,22 @@ def create_comment(request, article_pk):
 def comments_delete(request, article_pk, comment_pk):
     if request.user.is_authenticated:
         comment = get_object_or_404(Comment, pk=comment_pk)
-        comment.delete()
+        if comment.user == request.user:
+            comment.delete()
         return redirect('articles:detail', article_pk)    
     return HttpResponse('You are Unauthorized', status=401)
+
+def like(request, article_pk):
+    article = get_object_or_404(Article, pk=article_pk)
+    # 해당 게시글에 좋아요를 누른 사람들 중에서 현재 접속 유저가 있다면 좋아요를 취소
+    # 겟을 쓰면 오류가 생길수 있음
+    if article.like_users.filter(pk=request.user.pk).exists():
+        article.like_users.remove(request.user)
+    else:
+        article.like_users.add(request.user)
+
+    # if request.user in article.like_users.all():
+    #     article.like_users.remove(request.user) #좋아요 취소
+    # else:
+    #     article.like_users.add(request.user) #좋아요 
+    return redirect('articles:index')
